+++ packages/serinus/bin/serinus.dart
@@ -64,7 +64,7 @@ class GetRoute extends Route {
   int? get version => 2;
 
   @override
-  List<Guard> get guards => [TestGuard()];
+  List<Guard> get guards => [];
 }
 
 class PostRoute extends Route {
@@ -119,6 +119,9 @@ class AppModule extends Module {
   List<Pipe> get pipes => [
         TestPipe(),
       ];
+
+  @override
+  List<Guard> get guards => [TestGuard()];
 }
 
 class TestPipe extends Pipe {