+++ packages/serinus/lib/src/core/application.dart
@@ -1,14 +1,22 @@
 import 'dart:io';
 
 import 'package:meta/meta.dart';
-import 'package:serinus/src/commons/commons.dart';
-import 'package:serinus/src/commons/errors/initialization_error.dart';
-import 'package:serinus/src/commons/extensions/iterable_extansions.dart';
-import 'package:serinus/src/core/consumers/request_handler.dart';
-import 'package:serinus/src/core/containers/module_container.dart';
-import 'package:serinus/src/core/containers/router.dart';
-import 'package:serinus/src/core/core.dart';
-import 'package:serinus/src/core/injector/explorer.dart';
+
+import '../adapters/serinus_http_server.dart';
+import '../consumers/request_consumer.dart';
+import '../containers/module_container.dart';
+import '../containers/router.dart';
+import '../engines/view_engine.dart';
+import '../enums/enums.dart';
+import '../errors/initialization_error.dart';
+import '../extensions/iterable_extansions.dart';
+import '../global_prefix.dart';
+import '../http/http.dart';
+import '../injector/explorer.dart';
+import '../mixins/mixins.dart';
+import '../services/logger_service.dart';
+import '../versioning.dart';
+import 'core.dart';
 
 sealed class Application {
   final LogLevel level;
@@ -32,10 +40,6 @@ sealed class Application {
 
   String get url;
 
-  T? get<T extends Provider>() {
-    return modulesContainer.get<T>();
-  }
-
   HttpServer get server => config.serverAdapter.server;
 
   SerinusHttpServer get adapter => config.serverAdapter as SerinusHttpServer;
@@ -55,8 +59,6 @@ sealed class Application {
 
   @internal
   Future<void> shutdown();
-
-  Future<void> preview();
 
   Future<void> serve();
 
@@ -95,25 +97,14 @@ class SerinusApplication extends Application {
   }
 
   @override
-  Future<void> preview() async {
-    final explorer = Explorer(modulesContainer, router, config);
-    explorer.resolveRoutes();
-  }
-
-  @override
   Future<void> serve() async {
-    await preview();
+    await initialize();
     try {
-      _logger.info("Starting server on $url");
-      final handler = RequestHandler(router, modulesContainer, config);
+      _logger.info('Starting server on $url');
+      final handler = RequestConsumer(router, modulesContainer, config);
       await config.serverAdapter.listen(
-        (request, response) async {
-          await handler.handleRequest(request, response);
-        },
-        errorHandler: (e, stackTrace) {
-          print(e);
-          print(stackTrace);
-        },
+        handler.handleRequest,
+        errorHandler: (e, stackTrace) => _logger.severe(e, stackTrace),
       );
     } on SocketException catch (_) {
       _logger.severe('Failed to start server on $url');
@@ -134,7 +125,9 @@ class SerinusApplication extends Application {
           'The entry point of the application cannot be a DeferredModule');
     }
     await modulesContainer.registerModules(entrypoint, entrypoint.runtimeType);
-    await modulesContainer.finalize();
+    final explorer = Explorer(modulesContainer, router, config);
+    explorer.resolveRoutes();
+    await modulesContainer.finalize(entrypoint);
   }
 
   @override