+++ packages/serinus/lib/src/core/core.dart
@@ -1,6 +1,5 @@
 export 'application.dart';
 export 'application_config.dart';
-export 'contexts/contexts.dart';
 export 'controller.dart';
 export 'factory.dart' hide SerinusFactory;
 export 'guard.dart';