+++ packages/serinus/lib/src/core/guard.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'contexts/execution_context.dart';
+import '../contexts/execution_context.dart';
 
 abstract class Guard {
   const Guard();