+++ packages/serinus/lib/src/core/middleware.dart
@@ -1,4 +1,5 @@
-import 'package:serinus/serinus.dart';
+import '../contexts/request_context.dart';
+import '../http/http.dart';
 
 typedef NextFunction = Future<void> Function();