+++ packages/serinus/lib/src/enums/versioning_type.dart
@@ -1,0 +1,4 @@
+enum VersioningType {
+  header,
+  uri,
+}