+++ packages/serinus/lib/src/extensions/string_extensions.dart
@@ -1,0 +1,11 @@
+import 'dart:convert';
+
+extension JsonString on String {
+  dynamic tryParse() {
+    try {
+      return jsonDecode(this);
+    } catch (e) {
+      return null;
+    }
+  }
+}