+++ packages/serinus_cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.3
+
+- Re-add the dynamic fetch of the serinus version in the create command.
+
+## 0.2.2
+
+- Fix the create command.
+
 ## 0.2.1
 
 - Fix the create command to use the pre and post gen hooks.